feat: AskDeleteTimeModal - 코치일 때 상세 모달에서 띄워주기

+++ frontend/src/components/@common/InterviewDetailModal/index.tsx
@@ -3,14 +3,16 @@ import { useEffect, useState } from 'react';
 import * as S from './styled';
 
 import Accordion from '@/components/@common/Accordion';
+import AskDeleteTimeModal from '@/components/@common/AskDeleteTimeModal';
 import Modal from '@/components/@common/Modal';
+import useModal from '@/components/@common/Modal/useModal';
 
 import { useToastActions } from '@/context/ToastProvider';
 
 import { InterviewType } from '@/types/domain';
 import { MemberRole } from '@/types/domain';
 
-import { deleteCoachInterviewAPI, deleteCrewInterviewAPI, getInterviewAPI } from '@/api';
+import { deleteCrewInterviewAPI, getInterviewAPI } from '@/api';
 import { SUCCESS_MESSAGE } from '@/constants';
 import { getDateString, getTimeString } from '@/utils';
 
@@ -33,19 +35,22 @@ const InterviewDetailModal = ({
 }: InterviewDetailModalProps) => {
   const [interview, setInterview] = useState<InterviewType | null>();
 
+  const {
+    show: askShow,
+    display: askDisplay,
+    handleOpenModal: handleAskOpenModal,
+    handleCloseModal: handleAskCloseModal,
+  } = useModal();
+
   const { showToast } = useToastActions();
 
   const handleClickDeleteButton = async () => {
-    if (confirm('정말로 삭제하시겠습니까?')) {
-      if (role === 'CREW') {
-        await deleteCrewInterviewAPI(interviewId);
-        showToast('SUCCESS', SUCCESS_MESSAGE.CREW_DELETE_RESERVATION);
-      } else {
-        await deleteCoachInterviewAPI(interviewId);
-        showToast('SUCCESS', SUCCESS_MESSAGE.COACH_DELETE_RESERVATION);
-      }
-
+    if (role === 'CREW' && confirm('정말로 삭제하시겠습니까?')) {
+      await deleteCrewInterviewAPI(interviewId);
+      showToast('SUCCESS', SUCCESS_MESSAGE.CREW_DELETE_RESERVATION);
       afterDeleteInterview();
+    } else if (role === 'COACH') {
+      handleAskOpenModal();
     }
   };
 
@@ -116,6 +121,13 @@ const InterviewDetailModal = ({
           <Accordion key={question} title={question} description={answer} />
         ))}
       </S.AccordionContainer>
+      <AskDeleteTimeModal
+        show={askShow}
+        display={askDisplay}
+        interviewId={interviewId}
+        handleCloseModal={handleAskCloseModal}
+        afterDeleteInterview={afterDeleteInterview}
+      />
     </Modal>
   );
 };
